feat: aboutus.jsx style finished and code fixed
