+++ components/DeveloperCard/DeveloperCard.jsx
@@ -2,19 +2,10 @@ import styles from "./index.module.scss";
 import { useSelector } from "react-redux";
 import { useRouter } from "next/router";
 import Link from "next/link";
-import { AiFillGithub, AiFillLinkedin, AiFillInstagram } from "react-icons/ai";
-import { MdFacebook } from "react-icons/md";
+import { AiFillGithub, AiFillLinkedin } from "react-icons/ai";
 
 const DeveloperCard = ({ data }) => {
-  const {
-    personalImg,
-    name,
-    description,
-    fbprofile,
-    liprofile,
-    ghprofile,
-    igprofile,
-  } = data;
+  const { personalImg, name, description, liprofile, ghprofile } = data;
 
   return (
     <div className={styles.DeveloperCard}>
@@ -26,16 +17,6 @@ const DeveloperCard = ({ data }) => {
       <h1>{name}</h1>
       <p>{description}</p>
       <div className={styles.socials}>
-        <Link href={fbprofile} alt={name}>
-          <span alt="FacebookIcon">
-            <MdFacebook className={styles.icon} />
-          </span>
-        </Link>
-        <Link href={igprofile} alt={name}>
-          <span alt="InstagramIcon">
-            <AiFillInstagram className={styles.icon} />
-          </span>
-        </Link>
         <Link href={liprofile} alt={name}>
           <span alt="LinkedinIcon">
             <AiFillLinkedin className={styles.icon} />
