+++ components/DeveloperCard/index.module.scss
@@ -6,6 +6,7 @@
 
   color: rgba(255, 255, 255, 0.749);
   padding: 20px 60px;
+  height: 100%;
 
   .personalImg {
     height: 120px;
@@ -18,6 +19,7 @@
   }
 
   p {
+    font-size: 0.8rem;
   }
 
   .socials {
@@ -26,32 +28,44 @@
     .icon {
       height: 1.3rem;
       width: 1.3rem;
+      cursor: pointer;
+      transition: 0.45s;
+
+      &:hover {
+        color: white;
+      }
     }
   }
 }
 
 @media only screen and (min-width: 767px) {
   .DeveloperCard {
-    @include flex-container(column, center, center, $gap: 10px);
+    @include flex-container(column, center, center, $gap: 15px);
 
-    color: rgba(255, 255, 255, 0.749);
-    padding: 30px 60px;
+    padding: 30px 0px;
 
     .personalImg {
-      height: 120px;
-      width: 120px;
-      border-radius: 100px;
+      height: 160px;
+      width: 160px;
     }
 
     h1 {
       font-weight: bold;
+      font-size: 1.7rem;
     }
 
     p {
+      font-size: 1.2rem;
+      text-align: center;
     }
 
     .socials {
       @include flex-container($gap: 10px);
+
+      .icon {
+        height: 1.5rem;
+        width: 1.5rem;
+      }
     }
   }
 }
