+++ constants/index.js
@@ -1,1 +1,40 @@
 // export const pippo = "ciao";    ---sample of variables
+const developersData = [
+  {
+    personalImg: "https://picsum.photos/50/50",
+    name: "Noemi Fazio",
+    description: "Front-end Developer",
+    liprofile: "https://www.linkedin.com/in/noemi-fazio/",
+    ghprofile: "https://github.com/Waterlilyrr",
+  },
+  {
+    personalImg: "https://picsum.photos/50/50",
+    name: "Turi",
+    description: "Paninaro",
+    liprofile: "https://www.linkedin.com/in/noemi-fazio/",
+    ghprofile: "https://github.com/Waterlilyrr",
+  },
+  {
+    personalImg: "https://picsum.photos/50/50",
+    name: "Noemi Fazio",
+    description: "Front-end Developer",
+    liprofile: "https://www.linkedin.com/in/noemi-fazio/",
+    ghprofile: "https://github.com/Waterlilyrr",
+  },
+  {
+    personalImg: "https://picsum.photos/50/50",
+    name: "Turi",
+    description: "Paninaro",
+    liprofile: "https://www.linkedin.com/in/noemi-fazio/",
+    ghprofile: "https://github.com/Waterlilyrr",
+  },
+  {
+    personalImg: "https://picsum.photos/50/50",
+    name: "Noemi Fazio",
+    description: "Front-end Developer",
+    liprofile: "https://www.linkedin.com/in/noemi-fazio/",
+    ghprofile: "https://github.com/Waterlilyrr",
+  },
+];
+
+export default developersData;
